+++ sdk/storage/azure-storage-files-datalake/inc/azure/storage/files/datalake/protocol/datalake_rest_client.hpp
@@ -656,14 +656,14 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
           {
             request.AddHeader(
                 Details::HeaderIfModifiedSince,
-                createOptions.IfModifiedSince.GetValue().GetString(
+                createOptions.IfModifiedSince.GetValue().ToString(
                     Core::DateTime::DateFormat::Rfc1123));
           }
           if (createOptions.IfUnmodifiedSince.HasValue())
           {
             request.AddHeader(
                 Details::HeaderIfUnmodifiedSince,
-                createOptions.IfUnmodifiedSince.GetValue().GetString(
+                createOptions.IfUnmodifiedSince.GetValue().ToString(
                     Core::DateTime::DateFormat::Rfc1123));
           }
           if (createOptions.SourceIfMatch.HasValue())
@@ -679,14 +679,14 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
           {
             request.AddHeader(
                 Details::HeaderSourceIfModifiedSince,
-                createOptions.SourceIfModifiedSince.GetValue().GetString(
+                createOptions.SourceIfModifiedSince.GetValue().ToString(
                     Core::DateTime::DateFormat::Rfc1123));
           }
           if (createOptions.SourceIfUnmodifiedSince.HasValue())
           {
             request.AddHeader(
                 Details::HeaderSourceIfUnmodifiedSince,
-                createOptions.SourceIfUnmodifiedSince.GetValue().GetString(
+                createOptions.SourceIfUnmodifiedSince.GetValue().ToString(
                     Core::DateTime::DateFormat::Rfc1123));
           }
           return CreateParseResult(context, pipeline.Send(context, request));
@@ -752,14 +752,14 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
           {
             request.AddHeader(
                 Details::HeaderIfModifiedSince,
-                getPropertiesOptions.IfModifiedSince.GetValue().GetString(
+                getPropertiesOptions.IfModifiedSince.GetValue().ToString(
                     Core::DateTime::DateFormat::Rfc1123));
           }
           if (getPropertiesOptions.IfUnmodifiedSince.HasValue())
           {
             request.AddHeader(
                 Details::HeaderIfUnmodifiedSince,
-                getPropertiesOptions.IfUnmodifiedSince.GetValue().GetString(
+                getPropertiesOptions.IfUnmodifiedSince.GetValue().ToString(
                     Core::DateTime::DateFormat::Rfc1123));
           }
           return GetPropertiesParseResult(context, pipeline.Send(context, request));
@@ -823,14 +823,14 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
           {
             request.AddHeader(
                 Details::HeaderIfModifiedSince,
-                deleteOptions.IfModifiedSince.GetValue().GetString(
+                deleteOptions.IfModifiedSince.GetValue().ToString(
                     Core::DateTime::DateFormat::Rfc1123));
           }
           if (deleteOptions.IfUnmodifiedSince.HasValue())
           {
             request.AddHeader(
                 Details::HeaderIfUnmodifiedSince,
-                deleteOptions.IfUnmodifiedSince.GetValue().GetString(
+                deleteOptions.IfUnmodifiedSince.GetValue().ToString(
                     Core::DateTime::DateFormat::Rfc1123));
           }
           return DeleteParseResult(context, pipeline.Send(context, request));
@@ -901,14 +901,14 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
           {
             request.AddHeader(
                 Details::HeaderIfModifiedSince,
-                setAccessControlOptions.IfModifiedSince.GetValue().GetString(
+                setAccessControlOptions.IfModifiedSince.GetValue().ToString(
                     Core::DateTime::DateFormat::Rfc1123));
           }
           if (setAccessControlOptions.IfUnmodifiedSince.HasValue())
           {
             request.AddHeader(
                 Details::HeaderIfUnmodifiedSince,
-                setAccessControlOptions.IfUnmodifiedSince.GetValue().GetString(
+                setAccessControlOptions.IfUnmodifiedSince.GetValue().ToString(
                     Core::DateTime::DateFormat::Rfc1123));
           }
           request.AddHeader(Details::HeaderVersion, setAccessControlOptions.ApiVersionParameter);
@@ -1084,14 +1084,14 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
           {
             request.AddHeader(
                 Details::HeaderIfModifiedSince,
-                flushDataOptions.IfModifiedSince.GetValue().GetString(
+                flushDataOptions.IfModifiedSince.GetValue().ToString(
                     Core::DateTime::DateFormat::Rfc1123));
           }
           if (flushDataOptions.IfUnmodifiedSince.HasValue())
           {
             request.AddHeader(
                 Details::HeaderIfUnmodifiedSince,
-                flushDataOptions.IfUnmodifiedSince.GetValue().GetString(
+                flushDataOptions.IfUnmodifiedSince.GetValue().ToString(
                     Core::DateTime::DateFormat::Rfc1123));
           }
           request.AddHeader(Details::HeaderVersion, flushDataOptions.ApiVersionParameter);